bugfix: Fix behavior of edit contextmanager
The conditional for determining whether to create an empty Config object or to load in a pre-existing configuration was busted. If the pre-existing config existed, the context manager would return an empty Config object which is not the desired behavior. By adding `not` in front of the path check, the conditional now evaluates properly. Loads in the config if the config file exists; returns an empty Config object if it doesn't.

Also fixed a grammar error in the slurmconfig `_marshaller` function docstring.

Signed-off-by: Jason C. Nucciarone <jason.nucciarone@canonical.com>