chore(meta): Drop CLA check action
`slurmutils` has now moved to Charmed HPC, so it no longer needs the CLA check action to check if a contributing user is a member of the Canonical GitHub organisation. Libraries typically do not have a CLA check attached to it since they are not server-side or client-side applications where the license holder strongly matters.

Signed-off-by: Jason C. Nucciarone <nuccitheboss@ubuntu.com>