feat(editors): allow editors to set mode, user, and group of file
Adds the function `set_file_permissions` which is used by the dump function
of the editors to apply file mode, user, and group permissions to supported
configuration files. This makes Slurm happy because then we don't need to default
to root for everything, and can instead use the dedicated SlurmUser.

Signed-off-by: Jason C. Nucciarone <nuccitheboss@ubuntu.com>